Move common table to common folder
